Separate handler from server
