+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod admission_handler;
 pub mod patch;
 pub mod server;
 pub mod tls;
